+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Add support for newer ActiveJob features such as scheduled jobs and bulk enqueued jobs. https://github.com/Fullscript/yabeda-activejob/pull/24 @lewispb
+
 ## 0.6.0 - 2024-10-29
 - Add support for Rails 7.1.4+
 - Fix rubocop rules and configure Github actions to run the specs for supported Rails / Ruby versions.